fix Detail message changed to detail

+++ apps/docente_area/views.py
@@ -1,3 +1,5 @@
+import re
+
 from django.shortcuts import get_object_or_404
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
@@ -64,6 +66,8 @@ class DocenteAreaMixed(APIView):
     permission_classes = (IsAuthenticated,)
 
     def get(self, request, clave, value):
+        if re.search('[a-zA-Z]', value):
+            return Response(dict(detail=f'Error en valor: {value} al buscar {clave.split("_", 1)[0]}'))
         if clave == "docente_id":
             docenteAreas = DocenteArea.objects.filter(da_docente=value)
             if docenteAreas:
@@ -75,6 +79,8 @@ class DocenteAreaMixed(APIView):
             return Response(dict(detail="not found"))
 
     def put(self, request, clave, value):
+        if re.search('[a-zA-Z]', value):
+            return Response(dict(detail=f'Error en valor: {value} al buscar {clave.split("_", 1)[0]}'))
         if clave == "docente_id":
             DocenteArea.objects.filter(da_docente=value).delete()
             data = request.data.get("docenteArea")
